ref: SingleAssetPayments tests
